refactor(lambda): move users api endpoint into its owner lambda
