+++ aws/lambdas/strategies/main.py
@@ -41,7 +41,6 @@ def handler(event: Dict[str, Any], _context: Any) -> Dict[str, Any]:
     - GET /strategies/{id}/permissions
     - PATCH /strategies/{id}/permissions
     - DELETE /strategies/{id}/permissions
-    - GET /users/search
     """
     route_key = (event.get("requestContext") or {}).get("routeKey")
 
@@ -74,9 +73,6 @@ def handler(event: Dict[str, Any], _context: Any) -> Dict[str, Any]:
         strategy_id = (event.get("pathParameters") or {}).get("id")
         body = json.loads(event.get("body") or "{}")
         return update_permissions(strategy_id, body, event, mode="delete")
-
-    if route_key == "GET /users/search":
-        return search_users(event)
 
     return {"statusCode": 404, "body": "Not found"}
 
@@ -417,72 +413,6 @@ def _apply_permission_changes(
         return _json(500, {"message": "invalid mode"})
 
     return None
-
-
-def search_users(event: Dict[str, Any]) -> Dict[str, Any]:
-    netid = _get_netid_from_event(event)
-    if not netid:
-        return _json(401, {"message": "unauthorized"})
-
-    query_params = event.get("queryStringParameters") or {}
-    query = (query_params.get("q") or "").strip()
-    if len(query) < 2:
-        return _json(200, {"items": []})
-
-    limit_raw = query_params.get("limit")
-    try:
-        limit = int(limit_raw) if limit_raw is not None else 8
-    except ValueError:
-        limit = 8
-    limit = max(1, min(limit, 20))
-
-    query_lower = query.lower()
-    items: List[Dict[str, Any]] = []
-    seen = set()
-
-    def add_items(results: List[Dict[str, Any]]) -> None:
-        for item in results:
-            netid_value = str(item.get("netid") or "")
-            if not netid_value or netid_value in seen:
-                continue
-            seen.add(netid_value)
-            items.append(
-                {
-                    "netid": netid_value,
-                    "full_name": item.get("full_name") or "",
-                    "uconn_email": item.get("uconn_email") or "",
-                }
-            )
-
-    pk_value = "USER"
-
-    try:
-        netid_resp = USERS_TABLE.query(
-            IndexName="users-netid-gsi",
-            KeyConditionExpression=Key("search_pk").eq(pk_value)
-            & Key("netid").begins_with(query_lower),
-            ProjectionExpression="netid, full_name, uconn_email",
-            Limit=limit,
-        )
-        add_items(netid_resp.get("Items", []))
-    except Exception:
-        pass
-
-    if len(items) < limit:
-        remaining = limit - len(items)
-        try:
-            name_resp = USERS_TABLE.query(
-                IndexName="users-full-name-gsi",
-                KeyConditionExpression=Key("search_pk").eq(pk_value)
-                & Key("full_name_lower").begins_with(query_lower),
-                ProjectionExpression="netid, full_name, uconn_email",
-                Limit=remaining,
-            )
-            add_items(name_resp.get("Items", []))
-        except Exception:
-            pass
-
-    return _json(200, {"items": _clean_decimals(items)})
 
 
 # ----------------------------
